+++ packages/@stylexjs/babel-plugin/src/visitors/stylex-define-consts.js
@@ -8,6 +8,7 @@
  */
 
 import type { NodePath } from '@babel/traverse';
+import type { FunctionConfig } from '../utils/evaluate-path';
 
 import * as t from '@babel/types';
 import { evaluate } from '../utils/evaluate-path';
@@ -49,7 +50,18 @@ export default function transformStyleXDefineConsts(
     const args = callExpressionPath.get('arguments');
     const firstArg = args[0];
 
-    const { confident, value } = evaluate(firstArg, state);
+    const evaluatePathFnConfig: FunctionConfig = {
+      identifiers: {},
+      memberExpressions: {},
+      disableImports: true,
+    };
+    state.applyStylexEnv(evaluatePathFnConfig.identifiers);
+
+    const { confident, value } = evaluate(
+      firstArg,
+      state,
+      evaluatePathFnConfig,
+    );
     if (!confident) {
       throw callExpressionPath.buildCodeFrameError(
         messages.nonStaticValue('defineConsts'),