+++ packages/@stylexjs/babel-plugin/src/visitors/stylex-define-vars.js
@@ -126,6 +126,7 @@ export default function transformStyleXDefineVars(
       memberExpressions[name].positionTry = { fn: positionTry };
       identifiers[name] = { ...(identifiers[name] ?? {}), types: stylexTypes };
     });
+    state.applyStylexEnv(identifiers);
 
     const { confident, value } = evaluate(firstArg, state, {
       identifiers,