+++ packages/@stylexjs/stylex/src/stylex.js
@@ -157,6 +157,8 @@ export const when: StyleX$When = {
   },
 };
 
+export const env: $ReadOnly<{ [string]: mixed }> = Object.freeze({});
+
 export const types = {
   angle: <T: string | 0 = string | 0>(
     _v: ValueWithDefault<T>,
@@ -225,6 +227,7 @@ type IStyleX = {
   createTheme: StyleX$CreateTheme,
   defineConsts: StyleX$DefineConsts,
   defineVars: StyleX$DefineVars,
+  env: $ReadOnly<{ [string]: mixed }>,
   defaultMarker: () => $ReadOnly<{
     marker: 'default-marker',
     $$css: true,
@@ -272,5 +275,6 @@ _legacyMerge.props = props;
 _legacyMerge.types = types;
 _legacyMerge.when = when;
 _legacyMerge.viewTransitionClass = viewTransitionClass;
+_legacyMerge.env = env;
 
 export const legacyMerge: IStyleX = _legacyMerge;